changes to wishlist.js
