+++ client/src/components/WishListCard.js
@@ -3,49 +3,26 @@ import Card from 'react-bootstrap/Card'
 import CardGroup from 'react-bootstrap/CardGroup'
 
 function WishListCard(props) {
+  const title = 'Game Title'
+  const text = 'This game is heckin cool!'
+  const date = 'Release date'
   return (
     <CardGroup>
   <Card>
     <Card.Img variant="top" src="holder.js/100px160" />
     <Card.Body>
-      <Card.Title>Game Title</Card.Title>
+      <Card.Title>{title}</Card.Title>
       <Card.Text>
-       This game is heckin cool!
+       {text}
       </Card.Text>
     </Card.Body>
     <Card.Footer>
-      <small className="text-muted">Release Date!</small>
-    </Card.Footer>
-  </Card>
-  <Card>
-    <Card.Img variant="top" src="holder.js/100px160" />
-    <Card.Body>
-      <Card.Title>Game Title</Card.Title>
-      <Card.Text>
-      This game is heckin cool!
-      </Card.Text>
-    </Card.Body>
-    <Card.Footer>
-      <small className="text-muted">Release Date!</small>
-    </Card.Footer>
-  </Card>
-  <Card>
-    <Card.Img variant="top" src="holder.js/100px160" />
-    <Card.Body>
-      <Card.Title>Game Title</Card.Title>
-      <Card.Text>
-      This game is heckin cool!
-      </Card.Text>
-    </Card.Body>
-    <Card.Footer>
-      <small className="text-muted">Release Date!</small>
+      <small className="text-muted">{date}</small>
     </Card.Footer>
   </Card>
 </CardGroup>
-
   )
 }
-
 
 export default WishListCard;
