+++ package.json
@@ -19,6 +19,7 @@
     "apollo-server-express": "^2.15.0",
     "bootstrap": "^5.1.3",
     "express": "^4.17.1",
-    "mongoose": "^6.0.13"
+    "mongoose": "^6.0.13",
+    "react-router-dom": "^6.0.2"
   }
 }
